app/lib/ml/code-understanding/model/tf-model.ts
import * as tf from '@tensorflow/tfjs';
import type { TransformerLayerConfig, TrainingLogs, ModelCompileConfig, ModelPredictConfig } from './layers';

/**
 * TensorFlow.js Implementation of Code Understanding Transformer
 */

interface AttentionCache {
  keys: tf.Tensor;
  values: tf.Tensor;
  mask?: tf.Tensor;
}

interface ModelMetrics {
  loss: number;
  accuracy: number;
  perplexity: number;
  attentionScores: number[][];
}

export class TFTransformerModel {
  private config: TransformerLayerConfig;
  private model: tf.LayersModel | null;
  private optimizer: tf.Optimizer;
  private attentionCaches: Map<string, AttentionCache>;
  private metrics: ModelMetrics[];
  private initialized: boolean;
  private tokenEmbedding: tf.layers.Layer;
  private positionEmbedding: tf.layers.Layer;
  private encoderLayers: tf.layers.Layer[];
  private decoderLayers: tf.layers.Layer[];
  private outputLayer: tf.layers.Layer;
  private dropoutLayer: tf.layers.Layer;
  private layerNorm: tf.layers.Layer;

  constructor(config: TransformerLayerConfig) {
    this.config = config;
    this.model = null;
    this.initialized = false;
    this.attentionCaches = new Map();
    this.metrics = [];

    // Initialize layers
    this.tokenEmbedding = tf.layers.embedding({
      inputDim: config.vocabSize,
      outputDim: config.modelSize,
      embeddingsInitializer: 'glorotNormal',
      name: 'token_embedding'
    });

    this.positionEmbedding = tf.layers.embedding({
      inputDim: config.maxSequenceLength,
      outputDim: config.modelSize,
      embeddingsInitializer: 'glorotNormal',
      name: 'position_embedding'
    });

    this.dropoutLayer = tf.layers.dropout({
      rate: config.dropoutRate,
      name: 'embedding_dropout'
    });

    this.layerNorm = tf.layers.layerNormalization({
      epsilon: 1e-6,
      name: 'embedding_norm'
    });

    this.encoderLayers = [];
    this.decoderLayers = [];
    this.outputLayer = tf.layers.dense({
      units: config.vocabSize,
      activation: 'softmax',
      kernelInitializer: 'glorotNormal',
      name: 'output_layer'
    });

    // Initialize optimizer
    this.optimizer = tf.train.adam(config.learningRate);
  }

  /**
   * Initialize the model
   */
  async initialize(): Promise<void> {
    if (this.initialized) return;

    try {
      // Create encoder layers
      for (let i = 0; i < this.config.numLayers; i++) {
        const encoderLayer = this.createEncoderLayer(i);
        this.encoderLayers.push(encoderLayer);
      }

      // Create decoder layers
      for (let i = 0; i < this.config.numLayers; i++) {
        const decoderLayer = this.createDecoderLayer(i);
        this.decoderLayers.push(decoderLayer);
      }

      // Build the model
      const encoderInput = tf.input({
        shape: [this.config.maxSequenceLength],
        name: 'encoder_input'
      });
      const decoderInput = tf.input({
        shape: [this.config.maxSequenceLength],
        name: 'decoder_input'
      });

      // Encoder processing
      let encoderOutput = this.processEncoderInput(encoderInput);
      for (const encoderLayer of this.encoderLayers) {
        encoderOutput = encoderLayer.apply(encoderOutput) as tf.SymbolicTensor;
      }

      // Decoder processing
      let decoderOutput = this.processDecoderInput(decoderInput);
      for (const decoderLayer of this.decoderLayers) {
        decoderOutput = decoderLayer.apply([decoderOutput, encoderOutput]) as tf.SymbolicTensor;
      }

      // Final output processing
      const output = this.outputLayer.apply(decoderOutput) as tf.SymbolicTensor;

      // Create and compile the model
      this.model = tf.model({
        inputs: [encoderInput, decoderInput],
        outputs: output,
        name: 'code_transformer'
      });

      // Compile model
      this.model.compile({
        optimizer: this.optimizer,
        loss: tf.losses.softmaxCrossEntropy,
        metrics: ['accuracy']
      });

      this.initialized = true;
      console.log('TensorFlow.js model initialized');
      this.model.summary();

    } catch (error) {
      console.error('Error initializing TensorFlow.js model:', error);
      throw error;
    }
  }

  /**
   * Forward pass
   */
  async forward(input: number[][]): Promise<number[][]> {
    if (!this.initialized || !this.model) {
      throw new Error('Model not initialized');
    }

    const result = tf.tidy(() => {
      const inputTensor = tf.tensor2d(input);
      const output = this.model!.predict(inputTensor) as tf.Tensor;
      return output.arraySync() as number[][];
    });

    return result;
  }

  /**
   * Train the model
   */
  async train(
    inputs: number[][],
    targets: number[][],
    epochs: number = 1
  ): Promise<TrainingLogs> {
    if (!this.initialized || !this.model) {
      throw new Error('Model not initialized');
    }

    const inputTensor = tf.tensor2d(inputs);
    const targetTensor = tf.tensor2d(targets);

    const history = await this.model.fit(inputTensor, targetTensor, {
      epochs,
      batchSize: this.config.batchSize,
      validationSplit: 0.1,
      callbacks: {
        onEpochEnd: async (epoch, logs) => {
          console.log(`Epoch ${epoch + 1}: loss = ${logs?.loss}, accuracy = ${logs?.acc}`);
        }
      }
    });

    // Cleanup tensors
    inputTensor.dispose();
    targetTensor.dispose();

    return {
      loss: history.history['loss'] as number[],
      accuracy: history.history['acc'] as number[]
    };
  }

  /**
   * Save model
   */
  async saveModel(path: string): Promise<void> {
    if (!this.initialized || !this.model) {
      throw new Error('Model not initialized');
    }

    await this.model.save(`localstorage://${path}`);
  }

  /**
   * Load model
   */
  async loadModel(path: string): Promise<void> {
    try {
      this.model = await tf.loadLayersModel(`localstorage://${path}`);
      this.initialized = true;
    } catch (error) {
      console.error('Error loading model:', error);
      throw error;
    }
  }

  /**
   * Create encoder layer
   */
  private createEncoderLayer(index: number): tf.layers.Layer {
    const name = `encoder_layer_${index}`;

    // Multi-head attention
    const attention = tf.layers.dense({
      units: this.config.modelSize,
      name: `${name}_attention`
    });

    // Feed-forward network
    const ffn = tf.sequential({
      name: `${name}_ffn`,
      layers: [
        tf.layers.dense({
          units: this.config.feedForwardSize,
          activation: 'relu',
          name: `${name}_ffn_1`
        }),
        tf.layers.dense({
          units: this.config.modelSize,
          name: `${name}_ffn_2`
        })
      ]
    });

    // Layer normalization
    const layerNorm1 = tf.layers.layerNormalization({
      epsilon: 1e-6,
      name: `${name}_norm_1`
    });
    const layerNorm2 = tf.layers.layerNormalization({
      epsilon: 1e-6,
      name: `${name}_norm_2`
    });

    // Create encoder layer
    return tf.layers.rnn({
      cell: tf.layers.stackedRNNCells({
        cells: [
          tf.layers.simpleRNNCell({
            units: this.config.modelSize,
            name: `${name}_rnn`
          })
        ]
      }),
      returnSequences: true,
      name
    });
  }

  /**
   * Create decoder layer
   */
  private createDecoderLayer(index: number): tf.layers.Layer {
    const name = `decoder_layer_${index}`;

    // Similar to encoder layer but with cross-attention
    return tf.layers.rnn({
      cell: tf.layers.stackedRNNCells({
        cells: [
          tf.layers.simpleRNNCell({
            units: this.config.modelSize,
            name: `${name}_rnn`
          })
        ]
      }),
      returnSequences: true,
      name
    });
  }

  /**
   * Process encoder input
   */
  private processEncoderInput(input: tf.SymbolicTensor): tf.SymbolicTensor {
    // Token embeddings
    const embedded = this.tokenEmbedding.apply(input) as tf.SymbolicTensor;

    // Create position tensor and get position embeddings
    const positions = tf.range(0, this.config.maxSequenceLength).expandDims(0);
    const posEmbedded = this.positionEmbedding.apply(positions) as tf.SymbolicTensor;

    // Combine embeddings using a custom layer
    const addLayer = tf.layers.add({});
    const combined = addLayer.apply([embedded, posEmbedded]) as tf.SymbolicTensor;

    // Apply dropout and normalization
    const dropped = this.dropoutLayer.apply(combined) as tf.SymbolicTensor;
    return this.layerNorm.apply(dropped) as tf.SymbolicTensor;
  }

  /**
   * Process decoder input
   */
  private processDecoderInput(input: tf.SymbolicTensor): tf.SymbolicTensor {
    // Token embeddings
    const embedded = this.tokenEmbedding.apply(input) as tf.SymbolicTensor;

    // Create position tensor and get position embeddings
    const positions = tf.range(0, this.config.maxSequenceLength).expandDims(0);
    const posEmbedded = this.positionEmbedding.apply(positions) as tf.SymbolicTensor;

    // Combine embeddings using a custom layer
    const addLayer = tf.layers.add({});
    const combined = addLayer.apply([embedded, posEmbedded]) as tf.SymbolicTensor;

    // Apply dropout and normalization
    const dropped = this.dropoutLayer.apply(combined) as tf.SymbolicTensor;
    return this.layerNorm.apply(dropped) as tf.SymbolicTensor;
  }

  /**
   * Clean up resources
   */
  dispose(): void {
    if (this.model) {
      this.model.dispose();
      this.model = null;
    }

    // Clean up attention caches
    this.attentionCaches.forEach(cache => {
      cache.keys.dispose();
      cache.values.dispose();
      cache.mask?.dispose();
    });
    this.attentionCaches.clear();

    this.initialized = false;
  }
}

// Export factory function for creating model
export async function createTFModel(config: TransformerLayerConfig): Promise<TFTransformerModel> {
  const model = new TFTransformerModel(config);
  await model.initialize();
  return model;
}
